ai-server/requirements.txt
opencv-python>=4.12.0
ultralytics>=8.0.0
scipy>=1.9.0
fastdtw>=0.3.4
Pillow>=9.0.0
numpy>=2.0.0
gdown>=4.7.0

